+++ game-aplaguetalerequiem/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - None
 
-## [1.3.0] - 2025-11-11
+## [1.3.0] - 2025-11-14
 
 - Fixed game version detection for Xbox version.
+- Added buttons to open Config and Save folders (folder icon on Mods page toolbar).
+- Added custom launch tool.